refactor(@formatjs/editor): let Gazelle manage editor packages
