+++ codescythe.jsonc
@@ -13,7 +13,7 @@
     "packages/cli-lib/formatters/transifex.ts",
     "packages/ecma376/index.ts",
     "packages/editor/index.tsx",
-    "packages/editor/main.tsx",
+    "packages/editor/demo/main.tsx",
     "packages/editor/vrt/app.tsx",
     "packages/editor/vrt/*.config.ts",
     "packages/editor/vrt/server.ts",
