+++ knowledge-base/013-editor-vrt.md
@@ -14,7 +14,7 @@ flowchart LR
   Core --> Parser[ICU parser]
 ```
 
-`demo.tsx` is a separate StyleX consumer with localized, labeled native controls.
+`demo/demo.tsx` is a separate StyleX consumer with localized, labeled native controls.
 `design-system/` owns theme tokens and reusable controls. Vite compiles StyleX
 through `@stylexjs/unplugin` before the React plugin. Visual tests cover editing,
 invalid ICU with keyboard focus, and a narrow RTL layout.
@@ -50,8 +50,15 @@ Locale fallback and clamped pagination handle asynchronously changing inputs.
 categories inherit `other`, rather than multiplying a global placeholder count.
 
 `core.tsx` owns the original headless API; `index.tsx` exports it alongside the
-workflow and validation APIs. `demo.tsx` shares an optional StyleX `EditorView`
-with `workflow-demo.tsx`. Workflow state has no dependency on either UI module.
+workflow and validation APIs. `demo/demo.tsx` shares an optional StyleX `EditorView`
+with `demo/workflow-demo.tsx`. Workflow state has no dependency on either UI module.
 VRT's `?workflow=1` fixture exercises controlled persistence, filters, saved-state
 feedback, and desktop/narrow RTL screenshots. Existing demo baselines remain in
 place to catch regressions in the shared view.
+
+Gazelle maintains source and dependency lists in the headless, demo, and VRT
+Bazel packages. The demo is a separate internal library. VRT maps generated
+library and test rules to local `vrt_library` / `vrt_test` wrappers that preserve
+raw sources and typecheck them. Vite and Playwright own execution; the harness
+also checks the demo library and builds its server adapter.
+No editor package disables Gazelle.
