+++ packages/editor/demo/BUILD.bazel
@@ -1,0 +1,29 @@
+load("//tools:compile.bzl", "formatjs_library")
+load("//tools:tsconfig.bzl", "ESNEXT_TSCONFIG", "packages_tsconfig")
+
+package(default_visibility = ["//packages/editor:__subpackages__"])
+
+formatjs_library(
+    name = "demo",
+    srcs = [
+        "demo.tsx",
+        "design-system/components.tsx",
+        "design-system/tokens.stylex.ts",
+        "fixtures/en.json",  # keep
+        "fixtures/ru.json",  # keep
+        "main.tsx",
+        "workflow-demo.tsx",
+    ],
+    tsconfig = packages_tsconfig(ESNEXT_TSCONFIG),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//:node_modules/@stylexjs/stylex",
+        "//:node_modules/@types/react",
+        "//:node_modules/@types/react-dom",
+        "//:node_modules/lucide-react",
+        "//:node_modules/react",
+        "//:node_modules/react-dom",
+        "//:node_modules/react-intl",
+        "//packages/editor",
+    ],
+)
