+++ packages/editor/demo/fixtures/en-GB.json
