+++ packages/editor/demo/fixtures/en.json
